improve 64bit portability
for win64 mostly

+++ pico/memory.h
@@ -1,9 +1,11 @@
 // memory map related stuff
 
+#include "pico_port.h"
+
 typedef unsigned char  u8;
 typedef unsigned short u16;
 typedef unsigned int   u32;
-typedef unsigned long  uptr; // unsigned pointer-sized int
+typedef uintptr_t      uptr; // unsigned pointer-sized int
 
 #define M68K_MEM_SHIFT 16
 // minimum size we can map
